HAJ-1 int to long

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/UpdateStaffCmd.java
@@ -40,7 +40,7 @@ public class UpdateStaffCmd implements Serializable {
     @Length(max = 20)
     private String countryCode;
 
-    private int dateOfBirthHijri;
+    private Long dateOfBirthHijri;
 
     @Past
     private Date dateOfBirthGregorian;
